Math-Tutor-Backend/data-ingestion/requirements.txt
psycopg2-binary
pandas>=2.0.0,<3.0.0
sentence-transformers
numpy
datasets